+++ lib/src/data_class/to_map.dart
@@ -196,9 +196,9 @@ String _getToMapVarable(Varable v) {
 
     case EnumTypeVarable.list_data:
       if (yes_null_default_yes) {
-        return '$name?.map((e) => e?.toJson()).toList()  ?? $initComment';
+        return '$name?.map((e) => e.toJson()).toList()  ?? $initComment';
       } else if (yes_null_default_no) {
-        return '$name?.map((e) => e?.toJson()).toList()';
+        return '$name?.map((e) => e.toJson()).toList()';
       } else if (no_null_default_yes) {
         return '$name.map((e) => e.toJson()).toList()';
       } else if (no_null_default_no) {