+++ lib/src/enum/run_enum_string.dart
@@ -201,7 +201,6 @@ $fromValue
     }
   }
 
-  /// Pattern matching
   T map<T>({
 $mapStart
   }) {
@@ -210,7 +209,6 @@ $mapEnd
     }
   }
   
-  /// Pattern matching
   T maybeMap<T>({
     required T Function() orElse,
 $maybeMapStart
@@ -219,7 +217,6 @@ $maybeMapStart
 $maybeMapEnd
       );
 
-  /// Pattern matching
   T? maybeMapOrNull<T>({
 $maybeMapOrNullStart
   }) =>